SCOPED_TIMER reorganisation and cleanup
Now they're not named exactly according to class but according to hierarchy:
Every timer with the name "A::B" will be counted within timer "A" to make it easier to see who's part of what.

Unfortunately, there are still too many to be shown at the same time on the debug screen.

Also improved hash a bit to prevent collisions