Merge a vast amount of small changes from develop into maintenance This basically merges various changes that I don't think will be included with cherry-pick.
This has been merged manually by using a Folder diff and merge tool (meld) with very little typing.
All of these changes exist in the develop branch.
Most of the changes are obvious NLC (syntax cleanup, NULL->nullptr and similar).
In the case of more complex changes, those are bits of code contained in develop that I think should be included in maintenance and tested ASAP.

The purpose of this was to make the eventual switch to develop less open to (non-GL) bugs.
If there are any unwanted difference in behavior caused by this, please fix it through develop first.