Link libnghttp2 together with libcurl.
If cmake library searching was done correctly, libcurl would
automatically pull in all the required dependencies because they
are correctly specified in pkg-config etc, but currenly it's a
very static setup that doesn't permit that. Specifying manually
every single dependency of libcurl in the pr-downloader cmake
isn't the correct approach either, this seems like the least evil
one.