F16-F24 scancode keybind support (#305)
* F16-F24 scancode keybind support

These values aren't available to Lua, since SDL1 only supports up to F15. Also it appears that due to a bug in SDL2, F13+ will only work with scancodes and not keycodes at this time. 

Keycode pairs have been added for bindings anyway, for when the SDL-internal fix is released.

* Use loops for f1-f24 key/scancodes + changelog