Callouts: Add GetKeyFromScanSymbol(scanSymbol)
Returns the keysymbol correspondent to the keycode received by the game
when the scanCode represented by the scanSymbol is received. In other
terms: receives the position on the keyboard and returns the key in
spring keycode format (uikeys).

For example, on a qwertz keyboard (see https://kbdlayout.info/kbdgr
for reference):

Spring.GetKeyFromScanSymbol('sc_z') => 'y'
Spring.GetKeyFromScanSymbol('sc_[') => 'ü'