DualScreenMode: Use separate display for drawing
When viewport geometry is updated:

- Run through all displays currently connected
- Intersect their screen rectangles with the application windows
  rectangle
- Sort the intersections left to right and make positions relative to
  each other
- If DualScreenModeMiniMapOnLeft is enabled use the leftmost
  intersection as the dualView, otherwise use the rightmost one. Use the
  other screens as the main view
- Extract PosX, PosY, SizeX, SizeY and WindowOffsetY for dual and view from
  the intersections in the form of: viewPosX, dualViewPosX, ..., and so on

If view is offset from bottom of window it has a positive
PosY, same if offset from left of window for PosX.

If view is offset from top of the window it has a positive
WindowOffsetY.

For GL coordinates:

Origin starts at bottom left of main view and orients bottom-left to
top-right. GL viewport ensures this geometry is respected when drawing
to the screen. For dual mode rendering a temporary viewport is applied
internally so origin is bottom left of dualView with same orientation.

For in-game mouse coordinates:

Origin starts at bottom left of main view and orients bottom-left to
top-right. This implies negative or higher than main view geometry values
are possible when mouse coordinates refers to points in dual view.

For in-engine mouse coordinates:

Origin starts at left border of window and top border of main view and
orients top-left to bottom-right. All mouse coordinates are translated
before interacting or being passed to game space.

Also:

Fix cursor handling when leaving window