GetUnitEffectiveBuildRange accepts nil unitDefID
Some tasks (in particular, feature reclaim) do not have a relevant unitDefID.
In theory this means that the function should just accept a featureID, but I
don't want to do this before ticket #717 is done, and also because it needs
more design thought.

Some of the possible questions:
* what about terraform? How should the API look to request it?
* could featureDefID ever be a sensible thing to request?

Right now though, accepting `nil` acknowledges the existence of these
edge cases without actually addressing them in full.