consistently use `TeamHandler` for `AllyTeam` data
`AllyTeam` data is first loaded in `CGameSetup` from the start script,
and is later passed down to `TeamHandler` from `CGameSetup`.

Since `TeamHandler` is responsible for storing dynamic information
during the game, whereas `CGameSetup` only stores the initial data as
loaded from the start script, we remove all calls to
`CGameSetup::GetAllyStartingData` in favor of information coming from
`TeamHandler` to ensure that we consistently retrieve up-to-date data.