gamesetup: warn instead of fail when missing startpos in map configuration
Previously, we failed immediately when trying to start a game on a map
with insufficient startpos defined in its configuration to support the
number of teams specified in the start script. This behaviour was
denying the game a chance to handle start positions itself at a later
stage.

Now, we emit a warning instead when MapParser does not find startpos
defined in the map config for all teamIDs present in start script, but
continue to start the game normally anyway. This gives a chance to the
game to remedy the situation and handle startpos itself.

In the worst case, teams without a startpos will spawn in the corner --
giving an obvious hint that something is amiss and the map is unable to
support this number of players by itself without assistance from the
game.