Put cursornormal in springcontent
Currently the minimal Recoil game consists of two files:
* `modinfo.lua`
* `anims/cursornormal_0.bmp` or equivalent.

The former is fundamentally required to even be recognized
as a game; the latter is the only file that causes a crash
if missing. Everything else, even if technically needed, is
already provided by springcontent which cannot be opted out
of.

A default cursor is distributed within `cursors.sdz`, which
is not actually a mandatory dependency (like springcontent).

There is little reason not to provide the other cursors by
default as well, but that needs some design thought since
there are multiple approaches to that:
* become a dependency of springcontent, since there is
  already a precedent with spring bitmaps.
* become a mandatory dependency added by engine, same
  as springcontent.
* move all files to springcontent and leave cursors.sdz
  as just an empty husk.

Anyway, right now it's only a question of the one mandatory
file, so archives don't yet come into play.