Prevent rezbot from chasing unrepaired units in area resurrection (#2113)
The current behavior results in resurrected units within the resurrection area
being chased indefinitely by the rezbot, when they are not fully repaired.
This chase often leads to the rezbot reaching the front line, where it's often destroyed.

This patch modifies the behavior so that the rezbot stops repairing the resurrected unit
when it exits the direct repair range of the rezbot, i.e., when the unit move.
This solution is more prefered then repairing within resurrect area due
to the fact that resurrection often happen near the front line.